pyproject.toml
[project]
name = "nostr-shitposter-agent"
version = "0.1.0"
description = "A professional Nostr shitposter agent with AI capabilities"
requires-python = ">=3.10"
dependencies = [
    "agentstr-sdk>=0.6.15",
    "langchain-openai>=0.3.18",
    "python-dotenv>=1.1.0",
    "pydantic>=2.11.3",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.2.0",
    "pytest-asyncio>=1.0.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
asyncio_mode = "auto"
